requirements.txt
asgiref==3.8.1
attrs==24.2.0
awsebcli==3.21.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
boto3==1.35.71
botocore==1.35.71
cement==2.10.14
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==43.0.1
Deprecated==1.2.14
Django==5.1
django-crispy-forms==2.3
django-storages==1.14.4
django-tables2==2.7.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
h11==0.14.0
idna==3.8
jmespath==1.0.1
outcome==1.3.0.post0
packaging==24.2
paramiko==3.5.0
pathspec==0.10.1
psycopg2-binary==2.9.10
pycparser==2.22
PyGithub==2.4.0
PyJWT==2.4.0
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
selenium==4.24.0
semantic-version==2.10.0
sendgrid==6.11.0
setuptools==75.2.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
termcolor==2.5.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
whitenoise==6.8.2
wrapt==1.16.0
wsproto==1.2.0
